refactor most of the loaders into the postprocessing data interface
