+++ src/covid_model_seiir_pipeline/forecasting/data.py
@@ -57,7 +57,7 @@ class ForecastDataInterface:
                                name='theta')
         return thetas
 
-    def check_covariates(self, scenarios: Dict[str, ScenarioSpecification]) -> None:
+    def check_covariates(self, scenarios: Dict[str, ScenarioSpecification]) -> List[str]:
         regression_spec = io.load(self.regression_root.specification())
         # Bit of a hack.
         forecast_version = str(self.covariate_root._root)
@@ -85,7 +85,7 @@ class ForecastDataInterface:
                 if not io.exists(self.covariate_root[covariate](covariate_scenario=covariate_version)):
                     raise FileNotFoundError(f'No {covariate_version} file found for covariate {covariate}.')
 
-        return regression_covariates
+        return list(regression_covariates)
 
     def get_infectionator_metadata(self):
         regression_spec = io.load(self.regression_root.specification())
@@ -118,21 +118,9 @@ class ForecastDataInterface:
         data = pd.read_csv(data_path)
         return data.set_index('location_id')
 
-    def load_elastispliner_inputs(self):
-        metadata = self.get_infectionator_metadata()
-        deaths_version = Path(metadata['death']['metadata']['output_path'])
-        es_inputs = pd.read_csv(deaths_version / 'model_data.csv')
-        es_inputs['date'] = pd.to_datetime(es_inputs['date'])
-        return es_inputs
 
-    def load_elastispliner_outputs(self):
-        metadata = self.get_infectionator_metadata()
-        deaths_version = Path(metadata['death']['metadata']['output_path'])
-        noisy_outputs = pd.read_csv(deaths_version / 'model_results.csv')
-        noisy_outputs['date'] = pd.to_datetime(noisy_outputs['date'])
-        smoothed_outputs = pd.read_csv(deaths_version / 'model_results_refit.csv')
-        smoothed_outputs['date'] = pd.to_datetime(smoothed_outputs['date'])
-        return noisy_outputs, smoothed_outputs
+
+
 
     def load_total_deaths(self):
         """Load cumulative deaths by location."""
@@ -233,6 +221,9 @@ class ForecastDataInterface:
     def load_beta_params(self, draw_id: int) -> Dict[str, float]:
         df = io.load(self.regression_root.parameters(draw_id=draw_id))
         return df.set_index('params')['values'].to_dict()
+
+    def save_ode_params(self, ode_params: pd.DataFrame, scenario: str, draw_id: int):
+        io.dump(ode_params, self.forecast_root.ode_params(scenario=scenario, draw_id=draw_id))
 
     def save_components(self, forecasts: pd.DataFrame, scenario: str, draw_id: int):
         io.dump(forecasts, self.forecast_root.component_draws(scenario=scenario, draw_id=draw_id))
