+++ src/covid_model_seiir_pipeline/postprocessing/model/resampling.py
@@ -1,0 +1,41 @@
+"""Tools for resampling draw-level data."""
+from typing import Dict, List
+
+import numpy as np
+import pandas as pd
+
+
+def build_resampling_map(deaths, resampling_params):
+    cumulative_deaths = deaths.groupby(level='location_id').cumsum()
+    max_deaths = cumulative_deaths.groupby(level='location_id').max()
+    upper_deaths = max_deaths.quantile(resampling_params['upper_quantile'], axis=1)
+    lower_deaths = max_deaths.quantile(resampling_params['lower_quantile'], axis=1)
+    resample_map = {}
+    for location_id in max_deaths.index:
+        upper, lower = upper_deaths.at[location_id], lower_deaths.at[location_id]
+        loc_deaths = max_deaths.loc[location_id]
+        to_resample = loc_deaths[(upper < loc_deaths) | (loc_deaths < lower)].index.tolist()
+        np.random.seed(location_id)
+        to_fill = np.random.choice(loc_deaths.index.difference(to_resample),
+                                   len(to_resample), replace=False).tolist()
+        resample_map[location_id] = {'to_resample': to_resample,
+                                     'to_fill': to_fill}
+    return resample_map
+
+
+def resample_draws(measure_data: pd.DataFrame, resampling_map: Dict[int, Dict[str, List[int]]]):
+    output = []
+    locs = measure_data.reset_index().location_id.unique()
+    for location_id, loc_map in resampling_map.items():
+        if location_id not in locs:
+            continue
+        loc_data = measure_data.loc[location_id]
+        loc_data[loc_map['to_resample']] = loc_data[loc_map['to_fill']]
+        loc_data = pd.concat({location_id: loc_data}, names=['location_id'])
+        if isinstance(loc_data, pd.Series):
+            loc_data = loc_data.unstack()
+        output.append(loc_data)
+
+    resampled = pd.concat(output)
+    resampled.columns = [f'draw_{draw}' for draw in resampled.columns]
+    return resampled
