+++ src/covid_model_seiir_pipeline/postprocessing/task/postprocess.py
@@ -19,182 +19,6 @@ from covid_model_seiir_pipeline.forecasting.data import ForecastDataInterface
 from covid_model_seiir_pipeline.forecasting import postprocessing_lib as pp
 
 
-class MeasureConfig:
-    def __init__(self,
-                 loader: Callable[[str, ForecastDataInterface, int], Any],
-                 label: str,
-                 calculate_cumulative: bool = False,
-                 cumulative_label: str = None,
-                 aggregator: Callable = None):
-        self.loader = loader
-        self.label = label
-        self.calculate_cumulative = calculate_cumulative
-        self.cumulative_label = cumulative_label
-        self.aggregator = aggregator
-
-
-class CovariateConfig:
-    def __init__(self,
-                 loader: Callable[[str, bool, str, ForecastDataInterface, int], List[pd.Series]],
-                 label: str,
-                 time_varying: bool = False,
-                 draw_level: bool = False,
-                 aggregator: Callable = None):
-        self.loader = loader
-        self.label = label
-        self.time_varying = time_varying
-        self.draw_level = draw_level
-        self.aggregator = aggregator
-
-
-class OtherConfig:
-    def __init__(self,
-                 loader: Callable[[ForecastDataInterface], Any],
-                 label: str,
-                 is_table: bool = True,
-                 aggregator: Callable = None):
-        self.loader = loader
-        self.label = label
-        self.is_table = is_table
-        self.aggregator = aggregator
-
-
-MEASURES = {
-    'deaths': MeasureConfig(
-        pp.load_deaths,
-        'daily_deaths',
-        calculate_cumulative=True,
-        cumulative_label='cumulative_deaths',
-        aggregator=pp.sum_aggregator,
-    ),
-    'infections': MeasureConfig(
-        pp.load_infections,
-        'daily_infections',
-        calculate_cumulative=True,
-        cumulative_label='cumulative_infections',
-        aggregator=pp.sum_aggregator,
-    ),
-    'r_effective': MeasureConfig(
-        pp.load_r_effective,
-        'r_effective',
-    ),
-    'betas': MeasureConfig(
-        pp.load_betas,
-        'betas',
-    ),
-    'beta_residuals': MeasureConfig(
-        pp.load_beta_residuals,
-        'log_beta_residuals',
-    ),
-    'coefficients': MeasureConfig(
-        pp.load_coefficients,
-        'coefficients',
-    ),
-    'scaling_parameters': MeasureConfig(
-        pp.load_scaling_parameters,
-        'beta_scaling_parameters'
-    ),
-    'elastispliner_noisy': MeasureConfig(
-        pp.load_es_noisy,
-        'daily_elastispliner_noisy',
-        calculate_cumulative=True,
-        cumulative_label='cumulative_elastispliner_noisy',
-        aggregator=pp.sum_aggregator,
-    ),
-    'elastispliner_smoothed': MeasureConfig(
-        pp.load_es_smoothed,
-        'daily_elastispliner_smoothed',
-        calculate_cumulative=True,
-        cumulative_label='cumulative_elastispliner_smoothed',
-        aggregator=pp.sum_aggregator,
-    )
-}
-
-
-COVARIATES = {
-    'mobility': CovariateConfig(
-        pp.load_covariate,
-        'mobility',
-        time_varying=True,
-        draw_level=True,
-        aggregator=pp.mean_aggregator,
-    ),
-    'testing': CovariateConfig(
-        pp.load_covariate,
-        'testing',
-        time_varying=True,
-        aggregator=pp.mean_aggregator,
-    ),
-    'pneumonia': CovariateConfig(
-        pp.load_covariate,
-        'pneumonia',
-        time_varying=True,
-        # aggregator=pp.mean_aggregator,  This probably isn't what we want.  Still in discussion.
-    ),
-    'mask_use': CovariateConfig(
-        pp.load_covariate,
-        'mask_use',
-        time_varying=True,
-        aggregator=pp.mean_aggregator,
-    ),
-    'air_pollution_pm_2_5': CovariateConfig(
-        pp.load_covariate,
-        'air_pollution_pm_2_5',
-        aggregator=pp.mean_aggregator,
-    ),
-    'lri_mortality': CovariateConfig(
-        pp.load_covariate,
-        'lri_mortality',
-        aggregator=pp.mean_aggregator,
-    ),
-    'proportion_over_2_5k': CovariateConfig(
-        pp.load_covariate,
-        'proportion_over_2_5k',
-        aggregator=pp.mean_aggregator,
-    ),
-    'proportion_under_100m': CovariateConfig(
-        pp.load_covariate,
-        'proportion_under_100m',
-        aggregator=pp.mean_aggregator,
-    ),
-    'smoking_prevalence': CovariateConfig(
-        pp.load_covariate,
-        'smoking_prevalence',
-        aggregator=pp.mean_aggregator,
-    ),
-}
-
-MISCELLANEOUS = {
-    'full_data': OtherConfig(
-        pp.load_full_data,
-        'full_data',
-        aggregator=pp.sum_aggregator,
-    ),
-    'elastispliner_inputs': OtherConfig(
-        pp.load_elastispliner_inputs,
-        'full_data_es_processed',
-        aggregator=pp.sum_aggregator,
-    ),
-    'version_map': OtherConfig(
-        pp.build_version_map,
-        'version_map',
-    ),
-    'populations': OtherConfig(
-        pp.load_populations,
-        'populations',
-    ),
-    'hierarchy': OtherConfig(
-        pp.load_hierarchy,
-        'hierarchy',
-    ),
-    'locations_modeled_and_missing': OtherConfig(
-        pp.get_locations_modeled_and_missing,
-        'modeled_and_missing_locations',
-        is_table=False,
-    ),
-}
-
-
 def postprocess_measure(data_interface: ForecastDataInterface,
                         resampling_map: Dict[int, Dict[str, List[int]]],
                         scenario_name: str, measure: str,
