+++ src/covid_model_seiir_pipeline/postprocessing/task/resample_map.py
@@ -1,29 +1,30 @@
 from argparse import ArgumentParser, Namespace
-import logging
 from pathlib import Path
 import shlex
 from typing import Optional
 
+from covid_shared.cli_tools.logging import configure_logging_to_terminal
+from loguru import logger
 import pandas as pd
 
 from covid_model_seiir_pipeline import static_vars
-from covid_model_seiir_pipeline.forecasting.specification import ForecastSpecification, FORECAST_JOBS
-from covid_model_seiir_pipeline.forecasting.data import ForecastDataInterface
-from covid_model_seiir_pipeline.forecasting import postprocessing_lib as pp
-
-log = logging.getLogger(__name__)
+from covid_model_seiir_pipeline.postprocessing.data import PostprocessingDataInterface
+from covid_model_seiir_pipeline.postprocessing.specification import PostprocessingSpecification, POSTPROCESSING_JOBS
+from covid_model_seiir_pipeline.postprocessing.model import resampling, loaders
 
 
-def run_resample_map(forecast_version: str) -> None:
-    forecast_spec = ForecastSpecification.from_path(
-        Path(forecast_version) / static_vars.FORECAST_SPECIFICATION_FILE
+def run_resample_map(postprocessing_version: str) -> None:
+    postprocessing_spec = PostprocessingSpecification.from_path(
+        Path(postprocessing_version) / static_vars.POSTPROCESSING_SPECIFICATION_FILE
     )
-    workflow_spec = forecast_spec.workflow.task_specifications[FORECAST_JOBS.resample]
-    resampling_params = forecast_spec.resampling
-    data_interface = ForecastDataInterface.from_specification(forecast_spec)
-    deaths, *_ = pp.load_output_data(resampling_params.reference_scenario, data_interface, workflow_spec.num_cores)
+    workflow_spec = postprocessing_spec.workflow.task_specifications[POSTPROCESSING_JOBS.resample]
+    resampling_params = postprocessing_spec.resampling
+    data_interface = PostprocessingDataInterface.from_specification(postprocessing_spec)
+    deaths, *_ = loaders.load_output_data(resampling_params.reference_scenario,
+                                          data_interface,
+                                          workflow_spec.num_cores)
     deaths = pd.concat(deaths, axis=1)
-    resampling_map = pp.build_resampling_map(deaths, resampling_params)
+    resampling_map = resampling.build_resampling_map(deaths, resampling_params)
     data_interface.save_resampling_map(resampling_map)
 
 
@@ -31,9 +32,9 @@ def parse_arguments(argstr: Optional[str] = None) -> Namespace:
     """
     Gets arguments from the command line or a command line string.
     """
-    log.info("parsing arguments")
+    logger.info("parsing arguments")
     parser = ArgumentParser()
-    parser.add_argument("--forecast-version", type=str, required=True)
+    parser.add_argument("--postprocessing-version", type=str, required=True)
 
     if argstr is not None:
         arglist = shlex.split(argstr)
@@ -45,8 +46,10 @@ def parse_arguments(argstr: Optional[str] = None) -> Namespace:
 
 
 def main():
+    configure_logging_to_terminal(verbose=1)  # Debug level
+
     args = parse_arguments()
-    run_resample_map(forecast_version=args.forecast_version)
+    run_resample_map(postprocessing_version=args.postprocessing_version)
 
 
 if __name__ == '__main__':
